+++ src/test/java/org/realityforge/gwt/websockets/client/WebSocketTest.java
@@ -17,11 +17,14 @@ public class WebSocketTest
   public void registryTest()
   {
     assertNull( WebSocket.newWebSocketIfSupported() );
+    assertFalse( WebSocket.isSupported() );
     final TestWebSocket.Factory factory = new TestWebSocket.Factory();
     WebSocket.register( factory );
     assertNotNull( WebSocket.newWebSocketIfSupported() );
+    assertTrue( WebSocket.isSupported() );
     assertTrue( WebSocket.deregister( factory ) );
     assertNull( WebSocket.newWebSocketIfSupported() );
+    assertFalse( WebSocket.isSupported() );
     assertFalse( WebSocket.deregister( factory ) );
   }