+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.2:
 
+* Support WebSocket.isSupported() method that will return true if WebSocket.newWebSocketIfSupported()
+  will not return a null value.
 * Rename WebsocketEvent to WebSocketEvent to maintain consistent casing with the rest of the library.
 
 ## 0.1: